+++ settings.json
@@ -110,17 +110,22 @@
         }
     },
     "vehicle_robot": {
-        "name": "Robot vehicle",
+        "name": "Base robot vehicle",
         "settings": {
-            "direction_pins": [16,18],
-            "speed_pins": [24,26],
-            "speed_pwms": [0, 255],
-            "speeds": [0, 0.60],
             "diverged_speed": 0.30,
             "vehicle_delay": 0.05,
             "line_follower_delay": 0.01,
             "home_location": [0,0],
             "home_direction": 1
+        }
+    },
+    "vehicle_robot_raspberry_pi": {
+        "name": "Robot vehicle directly connected to a Raspberry Pi",
+        "settings": {
+            "direction_pins": [16,18],
+            "speed_pins": [24,26],
+            "speed_pwms": [0, 255],
+            "speeds": [0, 0.60]
         }
     },
     "reconstruction": {
